# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shorten urls for users",
        "title": "Scissor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "description": "Login with user details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "description": "Logout of user account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.LogoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/shortener/redirect/{shortURL}": {
            "get": {
                "description": "Redirect short url to original url",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL",
                        "name": "shortURL",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shortener/shorten": {
            "post": {
                "description": "Get original url and created a shortened version",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ShortenURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ShortenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shortener/url/{urlID}": {
            "delete": {
                "description": "Delete a url by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Url ID",
                        "name": "urlID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shortener/urls": {
            "get": {
                "description": "Retrieve all shortened urls created by each user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.GetURLsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Create a new user with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.SignupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Retrieve all existing user accounts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.GetUsersResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "description": "Retrieve an existing user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.GetUserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_0xMarvell_scissor_pkg_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.GetURLsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.URL"
                    }
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.GetUserResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "user": {}
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.GetUsersResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.User"
                    }
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "hero108"
                },
                "username": {
                    "type": "string",
                    "example": "linchung"
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "login successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.LogoutResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": " logout successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.ShortenResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "short url created successfully"
                },
                "short_url": {
                    "type": "string",
                    "example": "http://localhost:8080/api/v1/8KKoJCy"
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.ShortenURLRequest": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://google.com"
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.SignupRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "hero108"
                },
                "username": {
                    "type": "string",
                    "example": "linchung"
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.SignupResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "user": {
                    "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.User"
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.URL": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "original_url": {
                    "description": "The url to be shortened\n\nrequired: true\nexample: https://google.com/",
                    "type": "string"
                },
                "shortened_url": {
                    "description": "The shortened version of the original url\n\nrequired: false\nunique: true\nexample: djna7L",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "description": "The user ID of the user that created the shortened url\n\nexample: 1",
                    "type": "integer"
                }
            }
        },
        "github_com_0xMarvell_scissor_pkg_models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "urls": {
                    "description": "The urls created by the user\n\nrequired: false\nexample: password123",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_0xMarvell_scissor_pkg_models.URL"
                    }
                },
                "username": {
                    "description": "The username of the user\n\nrequired: true\nunique: true\nexample: john_doe",
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}